Move HiFiGANMelSpectrogram inside HiFiGANVocoderBundle
