+++ torchaudio/prototype/models/__init__.py
@@ -13,9 +13,7 @@ from .hifi_gan import (
     hifigan_generator_v1,
     hifigan_generator_v2,
     hifigan_generator_v3,
-    hifigan_mel_spectrogram,
     HiFiGANGenerator,
-    HiFiGANMelSpectrogram,
 )
 from .rnnt import conformer_rnnt_base, conformer_rnnt_model
