+++ torchaudio/prototype/pipelines/hifigan_pipeline.py
@@ -1,10 +1,13 @@
 from dataclasses import dataclass
-from typing import Any, Dict
+from typing import Any, Dict, Optional
 
+import torch
+import torch.nn.functional as F
 from torch.nn import Module
 from torchaudio._internal import load_state_dict_from_url
 
-from torchaudio.prototype.models.hifi_gan import hifigan_generator, HiFiGANGenerator, HiFiGANMelSpectrogram
+from torchaudio.prototype.models.hifi_gan import hifigan_generator, HiFiGANGenerator
+from torchaudio.transforms import MelSpectrogram
 
 
 @dataclass
@@ -21,11 +24,11 @@ class HiFiGANVocoderBundle:
     instances.
 
     This bundle can convert mel spectrorgam to waveforms and vice versa. A typical use case would be a flow like
-    `text -> mel spectrogram -> waveform`, where one can use an external component, e.g. Tactron2,
+    `text -> mel spectrogram -> waveform`, where one can use an external component, e.g. Tacotron2,
     to generate mel spectrogram from text. Please see below for the code example.
 
     .. note::
-        Although this works with the existing Tactron2 bundles, for the best results one needs to retrain Tactron2
+        Although this works with the existing Tacotron2 bundles, for the best results one needs to retrain Tacotron2
         using the same data preprocessing pipeline which was used for training HiFiGAN. In particular, the original
         HiFiGAN implemntation uses a custom method of generating mel spectrograms from waveforms, different from
         :py:class:`torchaudio.transforms.MelSpectrogram`. We reimplemented this transform as
@@ -55,12 +58,12 @@ class HiFiGANVocoderBundle:
         >>> # Since HiFiGAN bundle is in prototypes, it needs to be exported explicitly
         >>> from torchaudio.prototype.pipelines import HIFIGAN_VOCODER_V3_LJSPEECH as bundle_hifigan
         >>>
-        >>> # Load Tactron2 bundle
+        >>> # Load Tacotron2 bundle
         >>> bundle_tactron2 = torchaudio.pipelines.TACOTRON2_WAVERNN_CHAR_LJSPEECH
         >>> processor = bundle_tactron2.get_text_processor()
         >>> tacotron2 = bundle_tactron2.get_tacotron2()
         >>>
-        >>> # Use Tactron2 to convert text to mel spectrogram
+        >>> # Use Tacotron2 to convert text to mel spectrogram
         >>> text = "A quick brown fox jumped over a lazy dog"
         >>> input, lengths = processor(text)
         >>> specgram, lengths, _ = tacotron2.infer(input, lengths)
@@ -110,7 +113,7 @@ class HiFiGANVocoderBundle:
         `here
         <https://github.com/jik876/hifi-gan/blob/4769534d45265d52a904b850da5a622601885777/meldataset.py#L49-L72>`_.
         """
-        return HiFiGANMelSpectrogram(
+        return self._HiFiGANMelSpectrogram(
             n_mels=self._params["in_channels"],
             **self._mel_params,
         )
@@ -122,6 +125,74 @@ class HiFiGANVocoderBundle:
         :type: float
         """
         return self._mel_params["sample_rate"]
+
+    class _HiFiGANMelSpectrogram(torch.nn.Module):
+        """
+        Generate mel spectrogram in a way equivalent to the original HiFiGAN implementation:
+        https://github.com/jik876/hifi-gan/blob/4769534d45265d52a904b850da5a622601885777/meldataset.py#L49-L72
+
+        This class wraps around :py:class:`torchaudio.transforms.MelSpectrogram`, but performs extra steps to achive
+        equivalence with the HiFiGAN implementation.
+
+        Args:
+            hop_size (int): Length of hop between STFT windows.
+            n_fft (int): Size of FFT, creates ``n_fft // 2 + 1`` bins.
+            win_length (int): Window size.
+            f_min (float or None):  Minimum frequency.
+            f_max (float or None): Maximum frequency.
+            sample_rate (int):  Sample rate of audio signal.
+            n_mels (int):  Number of mel filterbanks.
+        """
+
+        def __init__(
+            self,
+            hop_size: int,
+            n_fft: int,
+            win_length: int,
+            f_min: Optional[float],
+            f_max: Optional[float],
+            sample_rate: float,
+            n_mels: int,
+        ):
+            super(HiFiGANVocoderBundle._HiFiGANMelSpectrogram, self).__init__()
+            self.mel_transform = MelSpectrogram(
+                sample_rate=sample_rate,
+                n_fft=n_fft,
+                win_length=win_length,
+                hop_length=hop_size,
+                f_min=f_min,
+                f_max=f_max,
+                n_mels=n_mels,
+                normalized=False,
+                pad=0,
+                mel_scale="slaney",
+                norm="slaney",
+                center=False,
+            )
+            self.sample_rate = sample_rate
+            self.hop_size = hop_size
+            self.n_fft = n_fft
+            self.win_length = win_length
+            self.f_min = f_min
+            self.f_max = f_max
+            self.n_mels = n_mels
+            self.pad_size = int((n_fft - hop_size) / 2)
+
+        def forward(self, waveform: torch.Tensor) -> torch.Tensor:
+            """Generate mel spectrogram from a waveform. Should have same sampling rate as ``self.sample_rate``.
+
+            Args:
+                waveform (Tensor): waveform of shape ``(batch_size, time_length)``.
+            Returns:
+                Tensor of shape ``(batch_size, n_mel, time_length)``
+            """
+            ref_waveform = F.pad(waveform.unsqueeze(1), (self.pad_size, self.pad_size), mode="reflect")
+            ref_waveform = ref_waveform.squeeze(1)
+
+            spectr = (self.mel_transform.spectrogram(ref_waveform) + 1e-9) ** 0.5
+            mel_spectrogram = self.mel_transform.mel_scale(spectr)
+            mel_spectrogram = torch.log(torch.clamp(mel_spectrogram, min=1e-5))
+            return mel_spectrogram
 
 
 HIFIGAN_VOCODER_V3_LJSPEECH = HiFiGANVocoderBundle(
